Running ok on linux

+++ runtest.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 # coding: utf-8
 
-import popen2, subprocess, socket, os, sys, re, platform, datetime, time, paramiko
+import subprocess, os, sys, re, platform, datetime, time, getpass
 from time import sleep
 
 so = platform.system()
@@ -22,7 +22,6 @@ class Component:
 			self.__conf2 = openconf(conf2)
 			self.__conf2place = conf2
 		self.identify_SO()			
-
 
 	def name(self):
         	return self.__name
@@ -46,7 +45,7 @@ class Component:
         	return self.__machine
 
 	def zipped(self):
-        	return self.__zipped_path.split(" ")[-1][:-2]
+        	return self.__zipped_path.split(sep)[-1]
 	
 	
 	def _get_time(self):
@@ -62,7 +61,6 @@ class Component:
                                          stdout=subprocess.PIPE,
                                          stderr=subprocess.STDOUT)
    		out, err = process.communicate()
-		print out
     		return out, err, process.returncode
 	"""
 	def execute_on_windows(self, remote_command, machine_addr, delay=None):
@@ -105,14 +103,14 @@ class Component:
 			self.execute("unzip -o "+zipped_path+" -d superzz", None)
 			self.execute("mkdir C:/beefs-tester/superzz/superzz", machine)
 			getdata_cmd = "sftp "+user+"@" + machine
-        		process = subprocess.Popen(getdata_cmd,
+        		process = subprocess.Popen(getdata_cmd,	
                                   shell=True,
 				   stdout=subprocess.PIPE,
                                          stderr=subprocess.STDOUT,stdin=subprocess.PIPE)
 			process.stdin.write("cd beefs-tester/superzz/\n")
 			process.stdin.write("put -r superzz/*\n")
 			process.stdin.write("quit\n")
-			print process.stdout.readline()
+			#print process.stdout.readline()
 			out, err = "","" #process.communicate()
         	return out, err, process.returncode
 
@@ -150,7 +148,7 @@ class Component:
         	                        src_path,
         	                        remote_path])
 			process = subprocess.Popen(getdata_cmd,
-                                  shell=True,
+                                   shell=True,
                                    stdout=subprocess.PIPE,
                                    stderr=subprocess.STDOUT)
         		out, err = process.communicate()
@@ -162,49 +160,53 @@ class Component:
 			process.stdin.write("cd "+dest_path+"\n")
 			process.stdin.write("put -r "+src_path+"\n")
 			process.stdin.write("quit\n")
-			out, err = "",""#process.communicate()
+			out, err = "","" 
         	return out, err, process.returncode
 
     	def mount(self, workload_path):
 		if self.so()=="Linux":
-			print self.machine()
-			print self._get_time()+"Cleaning previous files..."
-			self.execute("rm -r /tmp/superzz", self.machine())
-			self.execute("rm -r /tmp/"+self.zipped(), self.machine())
-			self.execute("rm -r /tmp/workloadzz", self.machine())
-			self.execute("rm -r /tmp/"+workload_path.split(sep)[-1], self.machine())
-			# Copy BeeFS zip
-			print self._get_time()+"Copying BeeFS ZIP ("+self.zipped()+") on "+ self.machine()
-			self.copy_zip(self.machine(), self.zipped())
-			# 	print self.execute("ls /tmp/", self.machine())
-			# Unzip the files of BeeFS
-			print self._get_time()+"Unzipping BeeFS ZIP: /tmp/superzz"
-			self.execute("unzip /tmp/"+self.zipped()+" -d /tmp/superzz", self.machine())
-			# Remove the zip, already unzipped
-			print self._get_time()+"Removing Zip..."
-			self.execute("rm /tmp/"+self.zipped(), self.machine())
-			if self.name() != "combee":
+			if self.name() in ["honeycomb","queenbee","combee"]:
+				print self._get_time()+"Cleaning previous files..."
+				
+				self.execute("rm -r /tmp/superzz", self.machine())
+				self.execute("rm -r /tmp/"+self.zipped(), self.machine())
+				self.execute("rm -r /tmp/workloadzz", self.machine())
+				self.execute("rm -r /tmp/"+workload_path.split(sep)[-1], self.machine())
+				# Copy BeeFS zip
+				print self._get_time()+"Copying BeeFS ZIP ("+self.zipped()+") on "+ self.machine()
+				self.copy_zip(self.machine(), self.zipped())
+				# print self.execute("ls /tmp/", self.machine())
+				# Unzip the files of BeeFS
+				print self._get_time()+"Unzipping BeeFS ZIP: /tmp/superzz"
+				self.execute("unzip /tmp/"+self.zipped()+" -d /tmp/superzz", self.machine())
+				# Remove the zip, already unzipped
+				print self._get_time()+"Removing Zip..."
+				#self.execute("rm /tmp/"+self.zipped(), self.machine())
+			if not self.name() == "combee":
 				# Copy the configuration file of component in remote machine that will run it
 				print self._get_time()+"Copying "+self.name()+".conf files into: tmp/superzz/conf/"+self.name()+".conf"
-				self.copy_files(self.machine(), self.conf_place(), "tmp/superzz/conf/"+self.name()+".conf")
+				self.copy_files(self.machine(), self.conf_place(), "tmp/superzz/conf/"+self.name()+".conf")	
+
 			else:
 				# Copy the configuration file of component in remote machine that will run it
 				print self._get_time()+"Copying honeycomb.conf files into: tmp/superzz/conf/honeycomb.conf"
 				self.copy_files(self.machine(), self.conf_place(), "tmp/superzz/conf/honeycomb.conf")
 				# Copy the configuration file of component in remote machine that will run it
 				print self._get_time()+"Copying honeybee.conf files into: tmp/superzz/conf/honeybee.conf"
-				self.copy_files(self.machine(), self.conf2_place(), "tmp/superzz/conf/honeybee.conf")
+				self.copy_files(self.machine(), self.conf2_place(), "tmp/superzz/conf/honeybee.conf") 
 			
 		
-			if self.name() is "honeycomb" or self.name() is "combee":
+			if self.name() == "honeycomb" or self.name() == "combee":
 				print self._get_time()+"Making directories needed to run BeeFS..."
-				self.execute("mkdir "+self.conf()["contributing_storage.directory"], self.machine())
-			if self.name() is "honeybee" or self.name() is "combee":
+				self.execute("mkdir "+self.conf()["contributing_storage.directory"], self.machine()) 
+			if self.name() == "honeybee" or self.name() == "combee":
 				print self._get_time()+"Making directories needed to run BeeFS..."
 				if self.name() is "combee":
 					self.execute("mkdir "+self.conf2()["mount_directory"], self.machine())
+					self.execute("rm -r "+self.conf2()["mount_directory"]+"/*", self.machine())
 				else:
 					self.execute("mkdir "+self.conf()["mount_directory"], self.machine())
+					self.execute("rm -r "+self.conf()["mount_directory"]+"/*", self.machine())
 				# Copy workload zip
 				print self._get_time()+"Copying workload.zip into /tmp"
 				self.copy_workload( self.machine(), workload_path)
@@ -266,14 +268,19 @@ class Component:
 	def identify_SO(self): #this method must be *runned* before all others
 		self.__so = self.execute("python -m platform", self.machine())[0].split("-")[0]
 
-	def kill_others(self):
+	def kill_others(self, component):
 			out, err, rcod = self.execute("ps xau | grep " + 
-                              	   	self.name().capitalize() + 
+                              	   	component + 
                                  	 " | grep "+user+"| grep -v grep", self.machine())
+			if "port 22: No route to host" in out:
+				print self._get_time()+self.machine()+" doesn't respond"
+				return False
       		        if out.split()!=[]:
 				print self._get_time()+"There is another instance of "+self.name()+" running on "+self.machine()
 				print self._get_time()+"Killing now... Process "+out.split()[1]
-				self.execute("kill "+out.split()[1], self.machine())
+				self.execute("kill "+out.split()[1], self.machine())	
+			return True
+			
 
 
 	def clear(self):
@@ -291,14 +298,71 @@ class Component:
 		self.execute("rm -r /tmp/workloadzz", self.machine())
 		self.execute("rm -r /tmp/workload.zip", self.machine())
 	
-	def start(self):
+	def test(self, workload_path):
+		# this script must be in beefs directory where exists \conf 
+		print self._get_time()+'Starting the test...Copying files into mount point'
+		if self.name() == "honeybee": 
+			honeybee = self.conf()
+			honeycomb = self.conf2()
+		else:	
+			honeybee = self.conf2()
+			honeycomb = self.conf()
+		if self.so() == "Linux":
+			source = "/tmp/workloadzz"
+			dest = honeybee['mount_directory']
+			command = "cp -r "+source+"/ "+dest+"/workload/" #"cp -r /tmp/workloadzz/ /tmp/beefs-mount/"
+		else: 
+			source = "C:\\beefs-tester\\workloadzz"
+			dest = 'Z:\ '
+			command = 'copy "'+source+'\*.*" "'+dest+'\\"'
+		queenbee = honeycomb['osdmaster'].split(':')
+		
+		psw = "09442518461j"#getpass.getpass("Password to "+user+"@"+machine())
+		
+		self.execute('ssh '+self.machine()+' "echo '+psw+' | sudo -S mkdir '+dest+'/workload/"', self.machine())
+		self.execute('ssh '+self.machine()+' "echo '+psw+' | sudo -S chown joopeeds '+dest+'/workload"', self.machine())
+		if queenbee[0]=="localhost":
+			allocation = "local"
+		else:
+			allocation = "distributed"
+		mode = honeycomb['file_synchronization'] +" , " + allocation + " and "+ self.name()
+		hostname = self.machine()
+		workload = workload_path.split(sep)[-1]
+		self.execute("unzip -o "+workload+" -d workloadzz", None)
+		size = getsize("workloadzz")
+		self.execute("rm -r workloadzz/", None)
+		dt = datetime.datetime.now()
+		test = ' Test on '+ self.so() +' '+ dt.strftime("%d/%m/%Y %H:%M:%S")
+		header = '#'*60 + '\n' + generateheader(test) + generateheader("Hostname of honeycomb: "+hostname) + generateheader("Queenbee on "+queenbee[0]) + generateheader("Mode: "+mode) + generateheader("Workload: "+ size)
+		work = workload.split('.')[0]
+		log = open('testlogs'+sep+work+sep+honeycomb['file_synchronization']+sep+allocation+sep+self.name()+sep+'test.log.'+dt.strftime("%d-%m-%Y.%Hh%Mmin%Ss")+'.log','w')
+		log.write(header)
+		startepoch = int(time.time())
+		self.execute(command, self.machine())
+		endepoch = int(time.time())
+		body =  'Copying from '+source+' to '+dest +'\n' + 'Command executed: '+command + '\n' + 'elapsed: ' + str( endepoch - startepoch )  +'s\n' + '#'*60 + '\n'
+		print self._get_time()+'Writting the log now...'
+		log.write(body)
+		log.close()
+		print self._get_time()+'ending the test now...'
+	
+
+	
+
+
+	def start(self, workload_path):
 		def componentisrunning(self):
 			sleep(5)
         		out, err, rcod = self.execute("ps xau | grep " + 
                               	   	self.name().capitalize() + 
                                  	 " | grep "+user+"| grep -v grep", self.machine())
       		        return out #if it is running, out is not empty, so it's is true
-
+		def componentismounted(self):
+			if self.name() == "honeybee": honeybee = self.conf()
+			else:honeybee = self.conf2()
+        		out, err, rcod = self.execute("df" + 
+                                 	 " | grep "+honeybee['mount_directory']+"| grep -v grep", self.machine())
+			return out #if it is running, out is not empty, so it's is true
 		# try to start component and return whether it is running or not
 		print self._get_time()+"Starting "+self.name()+" and waiting confirmation"
 		if self.so()=="Linux":
@@ -307,8 +371,17 @@ class Component:
 			self.start_on_windows("cd C:\\beefs-tester\\superzz\\", self.machine())
 		if componentisrunning(self):
 			print self._get_time()+self.name()+" is already running on "+self.machine()
+			if self.name()=="honeybee" or self.name()== "combee": 
+				if componentismounted(self):
+					print self._get_time()+"Mount point mounted!"
+					self.test(workload_path)
+				else:
+					print self._get_time()+"Mount point was not mounted, test couldn't be started"
+				#pass
 		else:
 			print self._get_time()+self.name()+" was NOT started on "+self.machine()
+			if self.name()=="honeybee" or self.name()== "combee": 
+				print "Test couldn't be started"
 
 	def stop(self):
 		def componentisrunning(self):
@@ -382,7 +455,7 @@ def opencomponentconf(file):
 
         config = []
         new = open(file).read().split("[========]")
-	zipped_path = new[0]
+	zipped_path = new[0].split()[2].split('\n')[0]
 	for i in range(len(new)):
 		config.append({})
 	for uni in range(1,len(new)):
@@ -393,12 +466,17 @@ def opencomponentconf(file):
 
 def getsize(source):
 	folder_size = 0
+	num_files = 0
 	for (path, dirs, files) in os.walk(source):
   		for file in files:
     			filename = os.path.join(path, file)
     			folder_size += os.path.getsize(filename)
-	return "%.1fMB" % (folder_size/(1024*1024.0))
+			num_files += 1
+	return "%.1fMB %d files" % (folder_size/(1024*1024.0), num_files)
 
+
+def get_time():
+		return int(datetime.datetime.now().strftime("%H")), int(datetime.datetime.now().strftime("%M"))
 
 def generateheader(text):
 	return  '#'*((60-len(text+'  '))/2) +' '+ text+' ' + '#'*((60-len(text+'  '))/2) +'\n'
@@ -406,7 +484,10 @@ def generateheader(text):
 
 def main(samples_config, zipped_path):
 	for sample_config in samples_config:
-		 if sample_config != {}:
+	    	while True: 
+			if get_time()[0] not in range(0,5)+range(18,24): break
+		sys.stdout.flush()
+		if sample_config != {}:
 			if not sample_config.has_key("combee"):
 		 	 samples = sample_config['samples']
 		 	 queenbee = sample_config['queenbee']
@@ -418,31 +499,31 @@ def main(samples_config, zipped_path):
 		 	 files_to_write = sample_config['files_to_write']
 		 	 data_server = Component("honeycomb", honeycomb, honeycomb_conf, "", zipped_path)
 		 	 meta_server = Component("queenbee", queenbee, queenbee_conf, "", zipped_path)
-		 	 client = Component("honeybee", honeybee, honeybee_conf, "", zipped_path)
+		 	 client = Component("honeybee", honeybee, honeybee_conf, honeycomb_conf, zipped_path)
 			 for i in range(int(sample_config['samples'])):
 				#print sample_config['files_to_write']
-				meta_server.identify_SO()
-				data_server.identify_SO()
-				client.identify_SO()
-				#FIXME
 				meta_server.clear()
 				data_server.clear()
 				client.clear()
-				meta_server.kill_others()
-				data_server.kill_others()
-				client.kill_others()
+				if not meta_server.kill_others("Queenbee"): continue
+				if not data_server.kill_others("Combee"): continue
+				if not data_server.kill_others("Honeycomb"): continue
+				if not client.kill_others("Honeybee"): continue
 				meta_server.mount(sample_config['files_to_write'])
 				data_server.mount(sample_config['files_to_write'])
 				client.mount(sample_config['files_to_write'])
-				meta_server.start()
-				data_server.start()
-				client.start()
-				#client.stop()
-				#data_server.stop()
-				#meta_server.stop()
-				#data_server.unmount()
-				#meta_server.unmount()
-				#client.unmount()
+				#sleep(180)
+				meta_server.start(sample_config['files_to_write'])
+				data_server.start(sample_config['files_to_write'])
+				client.start(sample_config['files_to_write'])
+				sys.stdout.flush()
+				#sleep(180)
+				client.stop()
+				data_server.stop()
+				meta_server.stop()
+				data_server.unmount()
+				meta_server.unmount()
+				client.unmount()
 			else:
 			 samples = sample_config['samples']
 		 	 queenbee = sample_config['queenbee']
@@ -454,57 +535,24 @@ def main(samples_config, zipped_path):
 		 	 combee = Component("combee", combee, honeycomb_conf, honeybee_conf, zipped_path)
 		 	 meta_server = Component("queenbee", queenbee, queenbee_conf, "", zipped_path)
 			 for i in range(int(sample_config['samples'])):
-				#print sample_config['files_to_write']
+				print sample_config['files_to_write']
 				meta_server.clear()
 				combee.clear()
-				meta_server.kill_others()
-				combee.kill_others()	
+				if not meta_server.kill_others("Queenbee"): continue
+				if not combee.kill_others("Combee"): continue
+				if not combee.kill_others("Honeycomb"): continue
+				if not combee.kill_others("Honeybee"): continue	
 				meta_server.mount(sample_config['files_to_write'])
 				combee.mount(sample_config['files_to_write'])
-				meta_server.start()
-				combee.start()
+				#sleep(180)
+				meta_server.start(sample_config['files_to_write'])
+				combee.start(sample_config['files_to_write'])
+				sys.stdout.flush()
+				#sleep(180)
 				combee.stop()
 				meta_server.stop()
 				meta_server.unmount()
 				combee.unmount()
-						
-			
-
-
-
-	"""
-	# this script must be in beefs directory where exists \conf  
-	honeybee = openconf('conf'+sep+'honeybee.conf')
-	honeycomb = openconf('conf'+sep+'honeycomb.conf')
-	source = sys.argv[1] # Type the origem without the last '\'
-	if so == "Linux":
-		dest = honeybee['mount_directory']
-		command = 'cp -r '+source+'/ '+dest+'/\n'
-	else: 
-		command = 'copy "'+source+'\*.*" "'+dest+'\\" \n'
-
-	queenbee = honeycomb['osdmaster'].split(':')
-	if queenbee[0]=="localhost":
-		allocation = "Non-Distributed"
-	else:
-		allocation = "Distributed"
-	mode = honeycomb['file_synchronization'] + allocation
-	hostname = socket.gethostname()
-	size = getsize(source)
-	queenbee = honeycomb['osdmaster'].split(':')
-
-	dt = datetime.datetime.now()
-	test = ' Test on '+ platform.system() +' '+ dt.strftime("%d/%m/%Y %H:%M ")
-	header = '#'*60 + '\n' + generateheader(test) + generateheader("Hostname of honeycomb: "+hostname) + generateheader("Queenbee on "+queenbee[0]) + generateheader("Mode: "+mode) + generateheader("Workload: "+ size)
-	log = open('testlogs'+sep+'test.log.'+dt.strftime("%d-%m-%Y.%Hh%M")+'.log','w')
-	log.write(header)
-	startepoch = int(time.time())
-	popen2.popen2(command)[0].readlines()
-	endepoch = int(time.time())
-	body =  'Copying from '+source+' to '+dest +'\n' + 'Command executed: '+command + '\n' + 'elapsed: ' + str( endepoch - startepoch )  +'s\n' + '#'*60 + '\n'
-	log.write(body)
-	log.close()
-	"""
 
 
 if __name__ == "__main__":
